move metering_sink
