+++ ydb/core/persqueue/events/internal.h
@@ -7,7 +7,7 @@
 #include <ydb/core/persqueue/blob.h>
 #include <ydb/core/persqueue/blob_refcounter.h>
 #include <ydb/core/persqueue/key.h>
-#include <ydb/core/persqueue/metering_sink.h>
+#include <ydb/core/persqueue/common/metering.h>
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
 #include <ydb/core/persqueue/percentile_counter.h>
 #include <ydb/core/persqueue/sourceid_info.h>
