+++ ydb/core/persqueue/pqtablet/metering_sink.h
@@ -5,17 +5,10 @@
 #include <util/generic/map.h>
 #include <util/generic/set.h>
 
+#include <ydb/core/persqueue/common/metering.h>
 
 namespace NKikimr::NPQ {
 
-enum class EMeteringJson {
-    PutEventsV1 = 1,
-    ResourcesReservedV1 = 2,
-    ThroughputV1 = 3,
-    ReadThroughputV1 = 4,
-    StorageV1 = 5,
-    UsedStorageV1 = 6,
-};
 
 class TMeteringSink {
 public:
