+++ ydb/core/persqueue/pqtablet/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    metering_sink.cpp
     pq_impl.cpp
     pq_impl_app.cpp
     pq_impl_app_sendreadset.cpp
