+++ ydb/core/persqueue/user_info.h
@@ -5,7 +5,7 @@
 #include "percentile_counter.h"
 #include "quota_tracker.h"
 #include "account_read_quoter.h"
-#include "metering_sink.h"
+#include <ydb/core/persqueue/pqtablet/metering_sink.h>
 #include "dread_cache_service/caching_service.h"
 
 #include <ydb/core/base/counters.h>
